v2: fix things
